refactor(skill): simplify update boundaries
